requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sentence-transformers==2.2.2
qdrant-client==1.6.9
ollama==0.1.7
pypdf2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
tavily-python==0.3.3
tiktoken>=0.5.2,<1.0
numpy==1.24.3
torch==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
PyJWT==2.8.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Frontend Dependencies
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
aiohttp==3.9.1
aiofiles==23.2.0
requests==2.32.4
asyncio-throttle==1.0.2
markdown==3.5.1
python-dateutil==2.8.2
keyring==24.3.0
cython==3.0.5
pyinstaller==6.2.0
pathlib2==2.3.7
uuid==1.30
